+++ plugins/chat/app/controllers/chat/api/channels_controller.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-CHANNEL_EDITABLE_PARAMS ||= %i[name description slug]
-CATEGORY_CHANNEL_EDITABLE_PARAMS ||= %i[
-  auto_join_users
-  allow_channel_wide_mentions
-  threading_enabled
-]
-
 class Chat::Api::ChannelsController < Chat::ApiController
+  CHANNEL_EDITABLE_PARAMS ||= %i[name description slug]
+  CATEGORY_CHANNEL_EDITABLE_PARAMS ||= %i[
+    auto_join_users
+    allow_channel_wide_mentions
+    threading_enabled
+  ]
+
   def index
     permitted = params.permit(:filter, :limit, :offset, :status)
