+++ plugins/chat/app/controllers/chat/api/channels_current_user_notifications_settings_controller.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-MEMBERSHIP_EDITABLE_PARAMS = %i[muted desktop_notification_level mobile_notification_level]
-
 class Chat::Api::ChannelsCurrentUserNotificationsSettingsController < Chat::Api::ChannelsController
+  MEMBERSHIP_EDITABLE_PARAMS = %i[muted desktop_notification_level mobile_notification_level]
+
   def update
     settings_params = params.require(:notifications_settings).permit(MEMBERSHIP_EDITABLE_PARAMS)
     membership_from_params.update!(settings_params.to_h)
