+++ plugins/chat/app/controllers/chat/api_controller.rb
@@ -2,17 +2,9 @@
 
 module Chat
   class ApiController < ::Chat::BaseController
-    before_action :ensure_logged_in
-    before_action :ensure_can_chat
-
     include Chat::WithServiceHelper
 
     private
-
-    def ensure_can_chat
-      raise Discourse::NotFound unless SiteSetting.chat_enabled
-      guardian.ensure_can_chat!
-    end
 
     def default_actions_for_service
       proc do
