+++ plugins/chat/app/controllers/chat/base_controller.rb
@@ -2,13 +2,14 @@
 
 module Chat
   class BaseController < ::ApplicationController
+    requires_plugin Chat::PLUGIN_NAME
+
     before_action :ensure_logged_in
     before_action :ensure_can_chat
 
     private
 
     def ensure_can_chat
-      raise Discourse::NotFound unless SiteSetting.chat_enabled
       guardian.ensure_can_chat!
     end
