+++ plugins/poll/lib/user_extension.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module DiscoursePoll
+  module UserExtension
+    extend ActiveSupport::Concern
+
+    prepended { has_many :poll_votes, dependent: :delete_all }
+  end
+end
